feat: refactor binary structure and expand library exports
Binary refactoring:
- Extract command routing to src/bin/cli/commands.rs
- Simplify src/bin/scope.rs to thin CLI wrapper
- Update scope-intercept.rs to use explicit imports

Library exports (src/lib.rs):
- Add comprehensive crate-level documentation
- Export CLI argument types (AnalyzeArgs, DoctorArgs, etc.)
- Export capture/logging utilities for CLI tools
- Export report builders and renderers
- Deprecate prelude module with migration guidance

Doctor module improvements:
- Add make_prompt_fn for bridging UserInteraction trait
- Update output format to use structured tracing fields
- Fix if-let chain pattern in cache migration
