+++ src/bin/cli/mod.rs
@@ -1,0 +1,3 @@
+//! CLI utilities and command routing for the scope binary.
+
+pub mod commands;
