+++ src/doctor/runner.rs
@@ -1,12 +1,12 @@
 use super::check::{ActionRunResult, ActionRunStatus, DoctorActionRun};
 use crate::doctor::check::RuntimeError;
+use crate::internal::prompts::UserInteraction;
 use crate::models::HelpMetadata;
-use crate::prelude::{
-    ActionReport, ActionTaskReport, CaptureOpts, ExecutionProvider, GroupReport, OutputDestination,
-    SkipSpec, generate_env_vars, progress_bar_without_pos,
+use crate::models::prelude::SkipSpec;
+use crate::shared::prelude::{
+    ActionReport, ActionTaskReport, CaptureOpts, DoctorGroup, ExecutionProvider, GroupReport,
+    OutputDestination, generate_env_vars, progress_bar_without_pos,
 };
-use crate::report_stdout;
-use crate::shared::prelude::DoctorGroup;
 use anyhow::Result;
 use colored::Colorize;
 use opentelemetry::trace::Status;
@@ -350,6 +350,10 @@ where
     }
 }
 
+/// Prompt the user for confirmation using the inquire crate.
+///
+/// This function wraps inquire::Confirm and handles TTY detection.
+/// It's used when `yolo` mode is disabled.
 fn prompt_user(prompt_text: &str, maybe_help_text: &Option<String>) -> bool {
     tracing_indicatif::suspend_tracing_indicatif(|| {
         let prompt = {
@@ -364,12 +368,31 @@ fn prompt_user(prompt_text: &str, maybe_help_text: &Option<String>) -> bool {
     })
 }
 
+/// Auto-approve all prompts.
+///
+/// This function automatically approves all prompts, used in `yolo` mode
+/// or when running non-interactively.
 fn auto_approve(prompt_text: &str, maybe_help_text: &Option<String>) -> bool {
-    println!("{} Yes (auto-approved)", prompt_text);
+    info!(target: "progress", prompt = %prompt_text, "Auto-approved");
     if let Some(help_text) = maybe_help_text {
-        println!("[{}]", help_text);
+        debug!(target: "progress", help = %help_text, "Auto-approve help text");
     }
     true
+}
+
+/// Create a prompt function from a UserInteraction implementation.
+///
+/// This bridges the gap between the trait-based UserInteraction interface
+/// and the function pointer interface used by DoctorActionRun.
+#[allow(dead_code)]
+pub fn make_prompt_fn<U: UserInteraction>(
+    user_interaction: &U,
+) -> impl Fn(&str, &Option<String>) -> bool + '_ {
+    move |prompt_text: &str, maybe_help_text: &Option<String>| {
+        tracing_indicatif::suspend_tracing_indicatif(|| {
+            user_interaction.confirm(prompt_text, maybe_help_text.as_deref())
+        })
+    }
 }
 
 async fn report_action_output<T>(
@@ -445,10 +468,8 @@ async fn print_pretty_result(
     let task_reports = action_task_reports_for_display(&result.action_report);
     for task in task_reports {
         if let Some(text) = task.output {
-            let line_prefix = format!("{group_name}/{action_name}");
             for line in text.lines() {
-                let output_line = format!("{}:  {}", line_prefix.dimmed(), line);
-                report_stdout!("{}", output_line);
+                error!(target: "user", group = group_name, action = action_name, "{}", line);
             }
         }
     }
@@ -527,7 +548,7 @@ mod tests {
     };
     use crate::doctor::runner::{GroupActionContainer, RunGroups, compute_group_order};
     use crate::doctor::tests::{group_noop, make_root_model_additional};
-    use crate::prelude::{ActionReport, ActionTaskReport, MockExecutionProvider};
+    use crate::shared::prelude::{ActionReport, ActionTaskReport, MockExecutionProvider};
     use anyhow::Result;
     use std::collections::{BTreeMap, BTreeSet};
     use std::sync::Arc;
